Feature
Added navigation with react router
and nested route inside contact form
as it depends on a specific route to be shown.

Form has been added but with default values just
so we can add a new burguer.

Orders page has been added to check all orders
have been done in the app.