Forms and Redux
Ended the forms section introducing
to patterns of validation and error handling
in forms. Max declares the state in the class
to know which input receive a determined validation
and so on.

Redux introduction with couple of examples of how
store, actions, reducers and action creators works.
Redux assignments to check if the concepts were actually
learned.