Redux integration burger App
Creating store, insertion of redux strategies
into react containers (where actually make sense),
by connecting with react-redux and using
redux state for handling variables when need
to pass those information to other components.

I tried to keep Max style of creating actions
and integrating but the way he uses the reducers
is certainly not my favorite so I decided to use
my own way.
I don't like the idea of loading a full module
in a file when you might actually need just a
couple of imports related to a file. I think
this is a great case for module desconstruction
to make our apps lighter.