Feature
Added generic validation for forms into
utility.js
Taking advantage of updateObject in
other parts of the app like in
the forms itself when updating the
data when the user inputs.

Removed console.logs through
the app when not needed.

Added lazy loading for the
different parts of the app.