Redux saga introduction
Here Max gives us a brief introduction but
very efficient of how redux-saga
uses es6 generators to take advantage
in middleware calls as an alternative
to redux-thunk. Not that
redux-thunk is bad, but redux-saga
takes it to the next level by
using this next level feature
of the language and of course
giving more power to the
developer when controlling
async code.

Topics covered:

* Redux saga setup
* Use of yield and delay
* substitution of redux-thunk by
sagas (functions*)